+++ src/ModuleManager.ts
@@ -156,14 +156,17 @@ export class ModuleManager {
 
 			if (currentModule) {
 				if (!currentModule._virtualDirname) {
-					throw ExceptionFactory.createAssertionError("g._require.resolve: couldn't resolve the moudle path without virtualPath");
+					throw ExceptionFactory.createAssertionError("g._require.resolve: couldn't resolve the module path without virtualPath");
 				}
 				resolvedPath = PathUtil.resolvePath(currentModule._virtualDirname, path);
 			} else {
-				if (!/^\.\//.test(path)) {
+				if (/^\.\//.test(path)) {
+					resolvedPath = path.substring(2);
+				} else if (/^\//.test(path)) {
+					resolvedPath = path.substring(1);
+				} else {
 					throw ExceptionFactory.createAssertionError("g._require.resolve: entry point path must start with './'");
 				}
-				resolvedPath = path.substring(2);
 			}
 
 			// 2.a. LOAD_AS_FILE(Y + X)