+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # ChangeLog
 
+## 3.20.0
+* @akashic/game-configuration@2.5.0 に追従
+  * `BundledTextAsset` を追加
+
+## 3.19.1
+* assetBundle が存在する状況においてスクリプトアセット以外を利用するとエラーとなる問題を修正
+
+## 3.19.0
+* @akashic/game-configuration@2.4.0 に追従
+  * `assetBundle` に対応
+  * `BundledScriptAsset` を追加
+
 ## 3.18.3
 * `g.E#destroy()` で破棄済みチェックを行うように修正